[WIP]: Moves default search fields.
