+++ api/lib/searchable-fields.ts
@@ -1,0 +1,30 @@
+export const defaultSearchableFields: string[] = [
+  'contents',
+  'title^2',
+  'publisher',
+  'author',
+  'title_alt',
+  'corp_author',
+  'editor',
+  'journal_title',
+  'essential_ocean_variables',
+  'sustainable_development_goals',
+  'refereed',
+  'abstract^2',
+  'publication_status',
+  'current_status',
+  'relation_uri',
+  'language',
+  'bptype',
+  'relation_is_part_of_series',
+  'type',
+  'subjects_other',
+  'subjects_instrument_type',
+  'subjects_parameter_discipline',
+  'subjects_dm_processes',
+  'identifier_orcid',
+  'identifier_doi',
+  'maturity_level',
+  'notes',
+  'coverage_spatial',
+];
